build script change

+++ makefile
@@ -82,6 +82,6 @@ build:all
 	grub-mkrescue -o okusha.iso okusha/
 	
 clean:
-	rm -f obj/*.o
-	rm -r -f okusha/
+	if [ -e obj ]; then rm -f obj/*.o; fi;
+	if [ -e okusha ]; then rm -r -f okusha/ ; fi;
 	if [ -e okusha.iso ]; then rm okusha.iso; fi;
